+++ app/src/main/java/com/example/ochev/callbacks/FontSizeListener.kt
@@ -1,5 +1,5 @@
 package com.example.ochev.callbacks
 
 fun interface FontSizeListener {
-    fun onInitialFontChanged(fontSize: Int)
+    fun onInitialFontChanged(fontSize: Float)
 }
